apps/api/pyproject.toml
[project]
name = "oralv-api"
version = "1.0.0"
description = "FastAPI backend for Oral Verification OS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.5",
  "argon2-cffi>=23.1.0",
  "celery[redis]>=5.6.0b2",
  "email-validator>=2.3.0",
  "fastapi>=0.135.3",
  "httpx>=0.28.1",
  "minio>=7.2.18",
  "pgvector>=0.4.1",
  "pydantic-settings>=2.11.0",
  "pypdf>=5.9.0",
  "psycopg[binary]>=3.2.10",
  "python-docx>=1.2.0",
  "python-multipart>=0.0.20",
  "python-pptx>=1.0.2",
  "redis>=6.4.0",
  "sqlalchemy>=2.0.49",
  "uvicorn[standard]>=0.37.0"
]

[dependency-groups]
dev = [
  "faker>=37.11.0",
  "mypy>=1.18.1",
  "pytest>=8.4.2",
  "pytest-cov>=7.0.0",
  "ruff>=0.13.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["oralv"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
plugins = []
